% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_jobs}
\alias{api_jobs}
\title{Manage API Jobs}
\usage{
api_jobs(api, req, res, job_id = NULL, subroute = NULL)
}
\arguments{
\item{api}{An object representing the API.}

\item{req}{The request object, indicating the method and potential body for creating or updating jobs.}

\item{res}{The response object.}

\item{job_id}{Optional; a string identifier for the job if specific job actions are required.}

\item{subroute}{Optional; specifies a particular job action such as 'estimate', 'logs', or 'results'.}
}
\value{
Depending on the request, this might return job details, creation confirmation, or the results of a specific job.
}
\description{
Handles the creation, updating, fetching, and deletion of jobs based on the request method and parameters.
}
