% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{atomic_save_rds}
\alias{atomic_save_rds}
\title{Atomically write an RDS object (temp file + rename)}
\usage{
atomic_save_rds(object, file, fail_msg = "Could not save RDS file")
}
\arguments{
\item{object}{R object to serialize.}

\item{file}{Destination path.}

\item{fail_msg}{Message passed to \code{\link[=api_stop]{api_stop()}} on failure.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Atomically write an RDS object (temp file + rename)
}
\keyword{internal}
