% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{empty_name}
\alias{empty_name}
\title{Sentinel formal used to distinguish missing arguments from defaults}
\usage{
empty_name()
}
\value{
The name object taken from a dummy function's formal \code{x}.
}
\description{
Sentinel formal used to distinguish missing arguments from defaults
}
\keyword{internal}
