% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{file_formats}
\alias{file_formats}
\title{Get Supported File Formats}
\usage{
file_formats()
}
\value{
A list containing two elements:
\describe{
\item{input}{A list of supported input formats.}
\item{output}{A list of supported output formats.}
}
}
\description{
This function returns a list of supported input and output file formats
for GIS data. Each format includes details such as title, description,
GIS data types, and parameters.
}
\examples{
formats <- file_formats()
names(formats$output)
}
