% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_namespace}
\alias{get_namespace}
\title{Get API process-graph evaluation namespace}
\usage{
get_namespace(api)
}
\arguments{
\item{api}{API object.}
}
\value{
An environment (\code{emptyenv()} parent) where process functions from
\code{processes_file} are defined and base operators are bound for sandboxed
\code{\link[=run_pgraph]{run_pgraph()}} evaluation.
}
\description{
Get API process-graph evaluation namespace
}
\keyword{internal}
