% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraphs.R
\name{is_pgraph}
\alias{is_pgraph}
\title{Test whether a list is a valid process graph wrapper}
\usage{
is_pgraph(pg)
}
\arguments{
\item{pg}{A list containing \code{process_graph}.}
}
\value{
Logical scalar: \code{TRUE} if every element is a node and exactly one
node is marked as the result.
}
\description{
Test whether a list is a valid process graph wrapper
}
\keyword{internal}
