% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraphs.R
\name{is_pnode}
\alias{is_pnode}
\title{Test whether a list is a single process node}
\usage{
is_pnode(node)
}
\arguments{
\item{node}{A list with optional names.}
}
\value{
Logical scalar: \code{TRUE} if \code{node} has \code{process_id} and \code{arguments} of
the expected types.
}
\description{
Test whether a list is a single process node
}
\keyword{internal}
