% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_cancel_results}
\alias{job_cancel_results}
\title{Cancel a running job and optionally clear result artefacts}
\usage{
job_cancel_results(api, user, job_id)
}
\arguments{
\item{api}{An openeocraft API object.}

\item{user}{The user identifier associated with the job.}

\item{job_id}{Identifier of the job to operate on.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Cancel a running job and optionally clear result artefacts
}
\keyword{internal}
