% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_populate_links}
\alias{job_populate_links}
\title{Attach openEO links to a job document}
\usage{
job_populate_links(job, api, req)
}
\arguments{
\item{job}{Job list (mutated links).}

\item{api}{API object.}

\item{req}{Plumber request.}
}
\value{
\code{job} with \code{self} and optional \code{results} links.
}
\description{
Attach openEO links to a job document
}
\keyword{internal}
