% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_reconcile_bg_process}
\alias{job_reconcile_bg_process}
\title{Mark running jobs as error when the callr worker has died.}
\usage{
job_reconcile_bg_process(api, user, job_id)
}
\description{
Mark running jobs as error when the callr worker has died.
}
\keyword{internal}
