% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_start}
\alias{job_start}
\title{Start a job asynchronously}
\usage{
job_start(api, user, job_id)
}
\arguments{
\item{job_id}{The identifier for the job.}
}
\value{
A list with a message and a status code.
}
\description{
This function starts a job based on the job_id provided. It checks for job validity,
updates job statuses, and processes the job asynchronously.
}
