% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_update}
\alias{job_update}
\title{Update job}
\usage{
job_update(api, user, job_id, job)
}
\arguments{
\item{job_id}{The identifier for the job}

\item{body}{The request body containing updates}
}
\description{
Update job
}
