% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraphs.R
\name{list_expr}
\alias{list_expr}
\title{Build a \code{list(...)} call from process arguments}
\usage{
list_expr(args, env, resolver = NULL)
}
\arguments{
\item{args}{List of argument values.}

\item{env}{Environment holding evaluated sibling nodes.}
}
\value{
A \code{call} object.
}
\description{
Build a \code{list(...)} call from process arguments
}
\keyword{internal}
