% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeocraft-package.R
\docType{package}
\name{openeocraft-package}
\alias{openeocraft}
\alias{openeocraft-package}
\title{openeocraft: Craft and Serve openEO API Applications in R}
\description{
Provides an R framework for crafting and serving openEO API applications by building on the plumber package. The friendly and extensible interface helps create RESTful openEO web services that prioritise ease of use, scalability, and flexibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Open-Earth-Monitor/openeocraft}
  \item Report bugs at \url{https://github.com/Open-Earth-Monitor/openeocraft/issues}
}

}
\author{
\strong{Maintainer}: Rolf Simoes \email{rolf.simoes@opengeohub.org} (\href{https://orcid.org/0000-0003-0953-4132}{ORCID})

Authors:
\itemize{
  \item Rolf Simoes \email{rolf.simoes@opengeohub.org} (\href{https://orcid.org/0000-0003-0953-4132}{ORCID})
  \item Brian Pondi \email{brian.pondi@uni-muenster.de} (\href{https://orcid.org/0009-0008-0367-1690}{ORCID})
}

Other contributors:
\itemize{
  \item OpenGeoHub Foundation \email{support@opengeohub.org} [copyright holder]
}

}
\keyword{internal}
