% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraphs.R
\name{pnode_args}
\alias{pnode_args}
\title{Convert each process argument to an R expression fragment}
\usage{
pnode_args(args, env, resolver = NULL)
}
\arguments{
\item{args}{Named list of arguments from one node.}

\item{env}{Environment of node id -> partially built expressions.}
}
\value{
List of language objects suitable for \code{as.call()}.
}
\description{
Convert each process argument to an R expression fragment
}
\keyword{internal}
