% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{process_decorators}
\alias{process_decorators}
\title{Scan a file for a given \verb{#* @decorator} tag and evaluate each chunk}
\usage{
process_decorators(api, file, decorator)
}
\arguments{
\item{api}{API object.}

\item{file}{Path to an R source file.}

\item{decorator}{Decorator name without \code{@} (e.g. \code{"openeo-process"}).}
}
\value{
\code{NULL}, invisibly.
}
\description{
Scan a file for a given \verb{#* @decorator} tag and evaluate each chunk
}
\keyword{internal}
