% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graphs.R
\name{process_graph_metadata}
\alias{process_graph_metadata}
\title{Metadata-only view of a UDP (omit large process_graph for list responses)}
\usage{
process_graph_metadata(udp)
}
\description{
Metadata-only view of a UDP (omit large process_graph for list responses)
}
\keyword{internal}
