% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_id}
\alias{random_id}
\title{Create a random identifier}
\usage{
random_id(n)
}
\arguments{
\item{n}{Length of the identifier in bytes.}
}
\value{
A hexadecimal string.
}
\description{
Create a random identifier
}
\examples{
id <- random_id(8)
nchar(id)
}
