% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraphs.R
\name{starting_pnode}
\alias{starting_pnode}
\title{Find the process node marked as the graph result}
\usage{
starting_pnode(pg)
}
\arguments{
\item{pg}{Validated process graph list.}
}
\value{
The result node list, or \code{NULL} if none found.
}
\description{
Find the process node marked as the graph result
}
\keyword{internal}
