% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{workspace_url_helpers}
\alias{workspace_url_helpers}
\alias{make_job_files_url}
\alias{make_workspace_files_url}
\title{Build signed URLs to artefacts in the workspace}
\usage{
make_job_files_url(host, user, job_id, file)

make_workspace_files_url(host, user, folder, file)
}
\arguments{
\item{host}{Service root URL.}

\item{user}{User identifier encoded into the token parameter.}

\item{job_id}{Identifier of the job whose files are requested.}

\item{file}{Relative file path inside the job folder.}

\item{folder}{Root-level folder under \verb{/files/root}.}
}
\value{
A character string containing the absolute URL.
}
\description{
Build signed URLs to artefacts in the workspace
}
