% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{format_helpers}
\alias{format_helpers}
\alias{format_ext}
\alias{ext_format}
\alias{ext_content_type}
\alias{format_content_type}
\title{Translate between file extensions, format identifiers, and MIME types}
\usage{
format_ext(format)

ext_format(filename)

ext_content_type(filename)

format_content_type(format)
}
\arguments{
\item{format}{A short format identifier such as \code{"gtiff"}.}

\item{filename}{A file name whose extension should be mapped.}
}
\value{
A character vector containing either the matching extension or
content type.
}
\description{
Translate between file extensions, format identifiers, and MIME types
}
