% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{make_job_files_url}
\alias{make_job_files_url}
\title{Build the download URL for a job artifact}
\usage{
make_job_files_url(host, user, job_id, file)
}
\arguments{
\item{host}{Service root URL.}

\item{user}{User identifier encoded into the token parameter.}

\item{job_id}{Identifier of the job whose files are requested.}

\item{file}{Relative file path inside the job folder.}
}
\value{
A character string containing the absolute URL.
}
\description{
Build the download URL for a job artifact
}
