% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraphs.R
\name{pgraph_fn}
\alias{pgraph_fn}
\title{Wrap \code{pgraph_expr()} as an executable function}
\usage{
pgraph_fn(pg, parent = NULL)
}
\arguments{
\item{pg}{Process graph list.}

\item{parent}{Passed to \code{pgraph_expr()}.}
}
\value{
A \code{function} with formals from \code{pgraph_params()} and body from
\code{pgraph_expr()}.
}
\description{
Wrap \code{pgraph_expr()} as an executable function
}
\keyword{internal}
