% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgraphs.R
\name{pnode_expr}
\alias{pnode_expr}
\title{Build a \code{call} for one process node}
\usage{
pnode_expr(node, env)
}
\arguments{
\item{node}{A process node list.}

\item{env}{Environment for resolving \code{from_node} references.}
}
\value{
A \code{call} object \code{process_id(...)}.
}
\description{
Build a \code{call} for one process node
}
\keyword{internal}
